chore[#50212]: Define `OpTurbo::{FrameComponent,StreamComponent}` remove wrapper
GitHub Discussion: https://github.com/opf/openproject/pull/13866#discussion_r1357813436

https://community.openproject.org/wp/50212
