+++ lookbook/previews/op_turbo/frame_component_preview/default.html.erb
@@ -1,0 +1,8 @@
+<p>
+  <strong>Default</strong><br/>
+  <%= render(OpTurbo::FrameComponent.new(model)) %>
+</p>
+<p>
+  <strong>Context</strong><br/>
+  <%= render(OpTurbo::FrameComponent.new(model, context: context)) %>
+</p>
