--- lookbook/previews/op_turbo/frame_wrapper_component_preview/default.html.erb
@@ -1,8 +1,0 @@
-<p>
-  <strong>Default</strong><br/>
-  <%= render(OpTurbo::FrameWrapperComponent.new(model)) %>
-</p>
-<p>
-  <strong>Context</strong><br/>
-  <%= render(OpTurbo::FrameWrapperComponent.new(model, context: context)) %>
-</p>
