+++ lookbook/previews/op_turbo/stream_component_preview/default.html.erb
@@ -1,0 +1,4 @@
+<p>
+  <strong>Default</strong><br/>
+  <%= render(OpTurbo::StreamComponent.new(template:, action:, target:)) %>
+</p>
