--- lookbook/previews/op_turbo/stream_wrapper_component_preview/default.html.erb
@@ -1,4 +1,0 @@
-<p>
-  <strong>Default</strong><br/>
-  <%= render(OpTurbo::StreamWrapperComponent.new(template:, action:, target:)) %>
-</p>
