+++ spec/components/op_turbo/frame_component_spec.rb
@@ -25,15 +25,27 @@
 #
 # See COPYRIGHT and LICENSE files for more details.
 #++
+#
+require 'spec_helper'
 
-module OpTurbo
-  class StreamWrapperComponent < ApplicationComponent
-    def initialize(template:, action:, target:)
-      super()
+RSpec.describe OpTurbo::FrameComponent, type: :component do
+  describe '#turbo_frame_id' do
+    context 'with `context:` option' do
+      it 'returns the turbo frame id' do
+        storage = build_stubbed(:nextcloud_storage, id: 1)
+        component = described_class.new(storage, context: :general_info)
 
-      @template = template
-      @action = action
-      @target = target
+        expect(component.turbo_frame_id).to eq('general_info_storages_nextcloud_storage_1')
+      end
+    end
+
+    context 'without `context:` option' do
+      it 'returns just the model dom id' do
+        storage = build_stubbed(:nextcloud_storage, id: 1)
+        component = described_class.new(storage)
+
+        expect(component.turbo_frame_id).to eq('storages_nextcloud_storage_1')
+      end
     end
   end
 end
